refactor: rename time-prefs plugin to os-prefs, add animator-duration-scale
time-prefs was about to gain a second, unrelated reading (Developer Options' animator duration scale) alongside its existing 12/24-hour time format read, so its name no longer matched its contents. Renamed to os-prefs, positioning it as this app's own counterpart to the official @tauri-apps/plugin-os (static device info) but for OS preferences instead -- a home for future reads of this kind rather than a new single-purpose plugin each time.
