+++ plugins/os-prefs/.gitignore
