+++ plugins/os-prefs/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "tauri-plugin-os-prefs"
+version = "0.1.0"
+authors = ["Threshold"]
+edition = "2021"
+links = "tauri-plugin-os-prefs"
+description = "Native OS preference reads for Threshold (time format, animation scale, ...)"
+
+[lib]
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[dependencies]
+tauri = { version = "2.0.0", features = [] }
+thiserror = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+[build-dependencies]
+tauri-plugin = { version = "2.0.0", features = ["build"] }
