+++ plugins/os-prefs/android/src/main/AndroidManifest.xml
