+++ plugins/os-prefs/package.json
@@ -1,0 +1,36 @@
+{
+	"name": "tauri-plugin-os-prefs-api",
+	"version": "0.1.0",
+	"author": "Threshold",
+	"description": "",
+	"type": "module",
+	"types": "./guest-js/index.ts",
+	"main": "./dist-js/index.cjs",
+	"module": "./dist-js/index.js",
+	"exports": {
+		"types": "./guest-js/index.ts",
+		"source": "./guest-js/index.ts",
+		"import": "./dist-js/index.js",
+		"require": "./dist-js/index.cjs",
+		"default": "./dist-js/index.js"
+	},
+	"files": [
+		"dist-js",
+		"README.md"
+	],
+	"scripts": {
+		"build": "rollup -c",
+		"prepublishOnly": "pnpm build",
+		"pretest": "pnpm build",
+		"typecheck": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@tauri-apps/api": "^2.0.0"
+	},
+	"devDependencies": {
+		"@rollup/plugin-typescript": "^12.0.0",
+		"rollup": "^4.9.6",
+		"typescript": "^5.3.3",
+		"tslib": "^2.6.2"
+	}
+}
