+++ plugins/os-prefs/permissions/autogenerated/commands/get_animator_duration_scale.toml
@@ -1,0 +1,13 @@
+# Automatically generated - DO NOT EDIT!
+
+"$schema" = "../../schemas/schema.json"
+
+[[permission]]
+identifier = "allow-get-animator-duration-scale"
+description = "Enables the get_animator_duration_scale command without any pre-configured scope."
+commands.allow = ["get_animator_duration_scale"]
+
+[[permission]]
+identifier = "deny-get-animator-duration-scale"
+description = "Denies the get_animator_duration_scale command without any pre-configured scope."
+commands.deny = ["get_animator_duration_scale"]
