+++ plugins/os-prefs/permissions/autogenerated/commands/get_time_format.toml
