+++ plugins/os-prefs/permissions/autogenerated/reference.md
@@ -1,0 +1,70 @@
+## Default Permission
+
+Default permissions for the os-prefs plugin
+
+#### This default permission set includes the following:
+
+- `allow-get-time-format`
+- `allow-get-animator-duration-scale`
+
+## Permission Table
+
+<table>
+<tr>
+<th>Identifier</th>
+<th>Description</th>
+</tr>
+
+
+<tr>
+<td>
+
+`os-prefs:allow-get-animator-duration-scale`
+
+</td>
+<td>
+
+Enables the get_animator_duration_scale command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`os-prefs:deny-get-animator-duration-scale`
+
+</td>
+<td>
+
+Denies the get_animator_duration_scale command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`os-prefs:allow-get-time-format`
+
+</td>
+<td>
+
+Enables the get_time_format command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`os-prefs:deny-get-time-format`
+
+</td>
+<td>
+
+Denies the get_time_format command without any pre-configured scope.
+
+</td>
+</tr>
+</table>
