+++ plugins/os-prefs/permissions/default.toml
@@ -1,0 +1,6 @@
+[default]
+description = "Default permissions for the os-prefs plugin"
+permissions = [
+    "allow-get-time-format",
+    "allow-get-animator-duration-scale",
+]
