+++ plugins/os-prefs/src/models.rs
@@ -1,0 +1,18 @@
+// Response models for OS preference queries
+//
+// (c) Copyright 2026 Liminal HQ, Scott Morris
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct TimeFormatResponse {
+    pub is24_hour: bool,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AnimatorDurationScaleResponse {
+    pub scale: f32,
+}
