--- plugins/time-prefs/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "tauri-plugin-time-prefs"
-version = "0.1.0"
-authors = ["Jules"]
-edition = "2021"
-links = "tauri-plugin-time-prefs"
-description = "Native Time Preferences for Threshold"
-
-[lib]
-crate-type = ["staticlib", "cdylib", "rlib"]
-
-[dependencies]
-tauri = { version = "2.0.0", features = [] }
-thiserror = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-[build-dependencies]
-tauri-plugin = { version = "2.0.0", features = ["build"] }
