--- plugins/time-prefs/permissions/autogenerated/commands/android.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-android"
-description = "Enables the android command without any pre-configured scope."
-commands.allow = ["android"]
-
-[[permission]]
-identifier = "deny-android"
-description = "Denies the android command without any pre-configured scope."
-commands.deny = ["android"]
