--- plugins/time-prefs/permissions/autogenerated/reference.md
@@ -1,69 +1,0 @@
-## Default Permission
-
-Default permissions for the time-prefs plugin
-
-#### This default permission set includes the following:
-
-- `allow-get-time-format`
-
-## Permission Table
-
-<table>
-<tr>
-<th>Identifier</th>
-<th>Description</th>
-</tr>
-
-
-<tr>
-<td>
-
-`time-prefs:allow-android`
-
-</td>
-<td>
-
-Enables the android command without any pre-configured scope.
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`time-prefs:deny-android`
-
-</td>
-<td>
-
-Denies the android command without any pre-configured scope.
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`time-prefs:allow-get-time-format`
-
-</td>
-<td>
-
-Enables the get_time_format command without any pre-configured scope.
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`time-prefs:deny-get-time-format`
-
-</td>
-<td>
-
-Denies the get_time_format command without any pre-configured scope.
-
-</td>
-</tr>
-</table>
