--- plugins/time-prefs/src/error.rs
@@ -1,26 +1,0 @@
-// Error type for time-prefs plugin operations
-//
-// (c) Copyright 2026 Liminal HQ, Scott Morris
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-use serde::{Serialize, Serializer};
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-    #[cfg(mobile)]
-    #[error("Mobile plugin not initialized")]
-    PluginInvoke(#[from] tauri::plugin::mobile::PluginInvokeError),
-}
-
-impl Serialize for Error {
-    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(self.to_string().as_ref())
-    }
-}
-
-pub type Result<T> = std::result::Result<T, Error>;
