--- plugins/time-prefs/src/models.rs
@@ -1,12 +1,0 @@
-// Response model for the system time-format query
-//
-// (c) Copyright 2026 Liminal HQ, Scott Morris
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TimeFormatResponse {
-    pub is24_hour: bool,
-}
